fix ci more
